- updated version to 0.2.6

+++ src/common/CXMapPainter2D.cpp
@@ -1179,7 +1179,7 @@ void CXMapPainter2D::OnInternalPaint(IBitmap *pBMP, IBitmap *pTmpBMP, int Width,
         TextRect.OffsetRect(0, bottom);
         pBMP->DrawTextASCII(ttt, TextRect, MAPFGCOLOR, MAPBGCOLOR);
         bottom = TextRect.GetBottom();
-        double FreeMemMB = 1.0*GetFreeMem()/1024/1024;
+        double FreeMemMB = 1.0*static_cast<int>(GetFreeMem()/1024/1024);
         ttt = CXStringASCII("MB: ") + FToA<CXStringASCII>(FreeMemMB, 1, 2);
         TextRect = pBMP->CalcTextRectASCII(ttt, 2, 2);
         TextRect.OffsetRect(0, bottom);
